refine code & docs
