+++ deploy/python_serving/pipeline_http_client.py
@@ -13,67 +13,11 @@
 # limitations under the License.
 
 import argparse
-import base64
 import json
-import os
-import os.path as osp
 
-import cv2
-import numpy as np
 import requests
 
-
-def numpy_to_base64(array: np.ndarray) -> str:
-    """numpy_to_base64
-
-    Args:
-        array (np.ndarray): input ndarray.
-
-    Returns:
-        bytes object: encoded str.
-    """
-    return base64.b64encode(array).decode('utf8')
-
-
-def video_to_numpy(file_path: str) -> np.ndarray:
-    """decode video with cv2 and return stacked frames
-       as numpy.
-
-    Args:
-        file_path (str): video file path.
-
-    Returns:
-        np.ndarray: [T,H,W,C] in uint8.
-    """
-    cap = cv2.VideoCapture(file_path)
-    videolen = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-    decoded_frames = []
-    for i in range(videolen):
-        ret, frame = cap.read()
-        # maybe first frame is empty
-        if ret is False:
-            continue
-        img = frame[:, :, ::-1]
-        decoded_frames.append(img)
-    decoded_frames = np.stack(decoded_frames, axis=0)
-    return decoded_frames
-
-
-def parse_file_paths(input_path: str) -> list:
-    assert osp.exists(input_path), \
-        f"{input_path} did not exists!"
-    if osp.isfile(input_path):
-        files = [
-            input_path,
-        ]
-    else:
-        files = os.listdir(input_path)
-        files = [
-            file for file in files
-            if (file.endswith(".avi") or file.endswith(".mp4"))
-        ]
-        files = [osp.join(input_path, file) for file in files]
-    return files
+from .utils import numpy_to_base64, parse_file_paths, video_to_numpy
 
 
 def parse_args():
